[update remove]メールアドレス変更の確認画面を削除 それに関与するルーティングも削除 メールアドレス変更後にログイン画面にリダイレクトするように変更
auth0のメタデータを反映させるために一度ログアウトさせる仕様に変更しました
