+++ frontend/app/src/components/model/profile/EmailChange.js
@@ -2,35 +2,58 @@ import axios from "axios";
 import { useForm } from "react-hook-form";
 import { useUser } from "../../../hooks/useUser";
 import ReactLoading from "react-loading";
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import form from "../../../css/ui/form.module.css";
 import { FcFeedback, FcHighPriority } from "react-icons/fc";
 import { AuthContext } from "../../../providers/AuthGuard";
-import { useNavigate } from "react-router-dom";
 
 const EmailChange = () => {
-  const navigate = useNavigate();
-  const { user, token } = useContext(AuthContext);
+  const { user, logout, loginWithRedirect } = useContext(AuthContext);
   const { useGetUser } = useUser();
   const { data: users, isLoading } = useGetUser();
+  const [ token, setToken ] = useState('');
+
+  const options = {
+    method: 'POST',
+    url: 'https://dev-mqrxqmfa.us.auth0.com/oauth/token',
+    data: {client_id: process.env.REACT_APP_AUTH0_MANAGEMENT_CLIENT_ID, client_secret : process.env.REACT_APP_AUTH0_MANAGEMENT_CLIENT_SECRET, audience: process.env.REACT_APP_AUTH0_MANAGEMENT_CLIENT_AUDIENCE, grant_type: process.env.REACT_APP_AUTH0_MANAGEMENT_GRANT_TYPE }
+  };
+
+  const getToken = async () => {
+    try {
+      const access_token = await axios(options);
+      setToken(access_token)
+    } catch (e) {
+      console.log(e.message)
+    }
+  };
+
+  useEffect(() => {
+    getToken();
+  }, []);
+
+  const onSubmit = async (data) => {
+    if (
+      window.confirm('メールアドレスを変更しますか？')
+    ) {
+      await axios.patch(`https://dev-mqrxqmfa.us.auth0.com/api/v2/users/${users.sub}`, data, {
+      headers: {
+        Authorization: `Bearer ${token.data.access_token}`,
+        'Content-Type': 'application/json',
+      },
+    })
+    .catch((error) => {
+      console.error(error.response.data);
+    });
+    alert('メールアドレスが変更されました!');
+    logout({ returnTo: window.location.origin })
+    loginWithRedirect({ redirect_url: `${window.location.origin}/mypage` })
+    }
+  };
 
   const { handleSubmit, register, formState: { errors } } = useForm({
     criteriaMode: "all"
   });
-
-  const onSubmit = async (data) => {
-    await axios.put(`${process.env.REACT_APP_DEV_API_URL}/user/users/${users.id}`, data, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-        'Content-Type': 'multipart/form-data',
-      },
-    })
-    .catch((error) => {
-      console.error(error.res.data);
-    });
-    navigate('/email-change-confirm');
-    window.location.reload();
-  }
 
   if(isLoading) return <ReactLoading type="spin" color='blue' className='loading' />
 
@@ -39,10 +62,7 @@ const EmailChange = () => {
       <form onSubmit={handleSubmit(onSubmit)} className={form.form}>
         <div className={form["form-text"]}>
           <div className={form["form-label"]}>現在のメールアドレス</div>
-          {users.e_mail === null && (
-            <div>{ user.email }</div>
-          )}
-          <div>{ users.e_mail }</div>
+          <div>{ user.email }</div>
         </div>
         <div className={form["form-text"]}>
           <div className={form["form-label"]}>変更するメールアドレス</div>
@@ -50,7 +70,7 @@ const EmailChange = () => {
           <input
           className={form["form-input"]}
           placeholder='メールアドレスを入力してください'
-          {...register('e_mail', {
+          {...register('email', {
             required: true
           })}
           />
