+++ frontend/app/src/route/Pages.js
@@ -22,7 +22,6 @@ export { default as GeneralUser } from '../components/model/user/GeneralUser';
 export { default as GeneralUserComic } from '../components/model/user/GeneralUserComic';
 export { default as ComicConfirmDelete } from '../components/model/mypage/ui/ComicConfirmDelete';
 export { default as ScenePostConfirmDelete } from '../components/model/scene_post/ui/ScenePostConfirmDelete';
-export { default as EmailChangeConfirm } from '../components/model/profile/ui/EmailChangeConfirm';
 export { default as ComicImageEdit } from '../components/model/mypage/ui/ComicImageEdit';
 export { default as ScenePostImageEdit } from '../components/model/scene_post/ui/ScenePostImageEdit';
 export { default as ProfileImageEdit } from '../components/model/profile/ui/ProfileImageEdit';
